Refactored liveblocks & server utilities paths
