+++ src/hooks.server.ts
@@ -1,5 +1,5 @@
 import type { Handle } from '@sveltejs/kit';
-import { auth } from '$lib/server/lucia';
+import { auth } from '$lib/server';
 
 export const handle = (async ({ event, resolve }) => {
 	event.locals.auth = auth.handleRequest(event);
