+++ src/utils/index.ts
@@ -1,9 +1,3 @@
-export * from './useMyPresence';
 export * from './createRoomId';
 export * from './str2color';
-export * from './useObject';
-export * from './useOthers';
 export * from './parseLang';
-export * from './useList';
-export * from './useSelf';
-export * from './useRoom';
